src/main/java/br/com/alura/dominio/aluno/Email.java
package br.com.alura.dominio.aluno;

public class Email {

    private String endereco;

    public Email(final String endereco) {
        String regexPattern = "^(?=.{1,64}@)[\\p{L}0-9_-]+(\\.[\\p{L}0-9_-]+)*@"
                + "[^-][\\p{L}0-9-]+(\\.[\\p{L}0-9-]+)*(\\.[\\p{L}]{2,})$";
        if (endereco == null || !endereco.matches(regexPattern)) {
            throw new IllegalArgumentException("E-mail incorreto");
        }
        this.endereco = endereco;
    }

    public String getEndereco() {
        return endereco;
    }

    public void setEndereco(String endereco) {
        this.endereco = endereco;
    }

}
